lint and reverse some renames
